+++ PongChat/pingpong/static/pingpong/js/script.js
@@ -10,22 +10,26 @@ import {
   drawRightPaddle,
   moveLeftPaddle,
   moveRightPaddle,
-  updateLeftPaddlePosition,
-  updatePaddleSize,
-  updatePaddleSpeed,
-  updateRightPaddlePosition,
+  updatePaddleElement,
 } from "./paddle.js";
 
 import {
   drawBall,
   moveBall,
-  updateBallPosition,
-  updateBallSize,
-  updateBallSpeed,
-  initializeBallPosition,
-  initializeBallSpeed,
+  initializeBallElement,
+  updateBallElement,
   ballY,
 } from "./ball.js";
+
+import {
+  startCountdown,
+  clearCanvas,
+  drawCenterLine,
+  drawCountdown,
+  drawScores,
+  countdownActive,
+  gamePaused,
+} from "./ui.js";
 
 // デュース機能を追加する
 //　点数のプログレスバーを追加する
@@ -38,39 +42,14 @@ import {
 // スタートボタンを作成する
 // スタートボタンを押すと、ゲームが始まる
 
-const canvas = document.getElementById("myCanvas");
-const ctx = canvas.getContext("2d");
-let previousCanvasWidth, previousCanvasHeight;
+export const canvas = document.getElementById("myCanvas");
+export const ctx = canvas.getContext("2d");
+export let previousCanvasWidth, previousCanvasHeight;
 
-// Score
-export let playerScore = 0;
-export let cpuScore = 0;
-export const winningScore = 11;
-
-// Countdown
-let countdown = 3;
-let countdownActive = true;
-let gamePaused = false;
-
-function initialize() {
+export function initialize() {
   updateCanvasSize();
-
-  updateBallSize(canvas);
-  initializeBallPosition(canvas);
-  initializeBallSpeed(canvas);
-
-  updatePaddleSize(canvas);
-  updateLeftPaddlePosition(previousCanvasHeight, canvas);
-  updateRightPaddlePosition(previousCanvasHeight, canvas);
-  updatePaddleSpeed(canvas);
-}
-
-export function incrementCpuScore() {
-  cpuScore++;
-}
-
-export function incrementPlayerScore() {
-  playerScore++;
+  initializeBallElement();
+  updatePaddleElement();
 }
 
 function updateCanvasSize() {
@@ -88,94 +67,6 @@ function updateCanvasSize() {
     canvas.width = windowWidth;
     canvas.height = windowWidth / aspectRatio;
   }
-}
-
-function drawCountdown() {
-  const fontSize = canvas.width * 0.2;
-  ctx.font = `${fontSize}px Arial`;
-  ctx.fillStyle = "rgba(0, 149, 221, 0.5)";
-
-  ctx.textAlign = "center";
-  ctx.textBaseline = "middle";
-  if (countdown === 0) {
-    ctx.fillText("Start!", canvas.width / 2, canvas.height / 2);
-  } else {
-    ctx.fillText(countdown, canvas.width / 2, canvas.height / 2);
-  }
-}
-
-function drawCenterLine() {
-  const lineWidth = canvas.width * 0.01;
-  const radius = lineWidth * 0.7;
-
-  ctx.beginPath();
-  ctx.setLineDash([]);
-  ctx.fillStyle = "#0095DD";
-
-  for (let y = radius; y < canvas.height; y += radius * 4) {
-    ctx.moveTo(canvas.width / 2, y);
-    ctx.arc(canvas.width / 2, y, radius, 0, Math.PI * 2);
-  }
-
-  ctx.fill();
-  ctx.closePath();
-}
-
-function drawScores() {
-  const fontSize = canvas.width * 0.2;
-  ctx.font = `${fontSize}px Arial`;
-  ctx.fillStyle = "rgba(0, 149, 221, 0.5)";
-
-  ctx.textAlign = "center";
-  ctx.textBaseline = "middle";
-  ctx.fillText(playerScore, canvas.width / 4, canvas.height / 2);
-  ctx.fillText(cpuScore, (canvas.width / 4) * 3, canvas.height / 2);
-}
-
-function clearCanvas() {
-  ctx.clearRect(0, 0, canvas.width, canvas.height);
-}
-
-function startCountdown() {
-  const interval = setInterval(() => {
-    if (countdown === 0) {
-      countdownActive = false;
-      clearInterval(interval);
-    } else {
-      countdown--;
-    }
-  }, 1000);
-}
-
-function showRestartButton(message) {
-  const gameOverMessage = document.getElementById("game-over-message");
-  const gameOverText = document.getElementById("gameOverText");
-  const restartButton = document.getElementById("restartButton");
-
-  gameOverText.textContent = message;
-  gameOverMessage.classList.remove("d-none");
-
-  restartButton.onclick = () => {
-    playerScore = 0;
-    cpuScore = 0;
-    gameOverMessage.classList.add("d-none");
-    initialize();
-    startCountdown();
-    gamePaused = false;
-  };
-}
-
-export function gameOver(message) {
-  gamePaused = true;
-  showRestartButton(message);
-}
-
-export function resetGame() {
-  gamePaused = true;
-  setTimeout(() => {
-    initialize();
-    gamePaused = false;
-  }, 500);
 }
 
 function draw() {
@@ -207,15 +98,8 @@ function onResize() {
   previousCanvasHeight = canvas.height;
 
   updateCanvasSize();
-
-  updateBallSize(canvas);
-  updateBallPosition(canvas, previousCanvasWidth, previousCanvasHeight);
-  updateBallSpeed(canvas);
-
-  updatePaddleSize(canvas);
-  updateLeftPaddlePosition(previousCanvasHeight, canvas);
-  updateRightPaddlePosition(previousCanvasHeight, canvas);
-  updatePaddleSpeed(canvas);
+  updateBallElement();
+  updatePaddleElement();
 }
 
 initialize();
@@ -223,8 +107,6 @@ startCountdown();
 
 // resize event
 window.addEventListener("resize", onResize, false);
-
-// keybord event
 document.addEventListener("keydown", keyDownHandler, false);
 document.addEventListener("keyup", keyUpHandler, false);
